Renamed oneof to oneOf (and moved to helpers dir)
