+++ src/index.js
@@ -1,4 +1,4 @@
 export { default as compose } from './compose'
 export { default as createSpec } from './createSpec'
-export { default as oneof } from './oneof'
 export { default as faker } from './faker'
+export * from './helpers'
